Describe your changes here

+++ db/migrate/20240807085618_drop_tasks.rb
@@ -1,9 +1,21 @@
 class DropTasks < ActiveRecord::Migration[7.1]
   def up
-    drop_table :tasks
+    # Check if the table exists before trying to drop it
+    if table_exists?(:tasks)
+      drop_table :tasks
+    else
+      puts "Table 'tasks' does not exist, skipping drop."
+    end
   end
 
   def down
-    raise ActiveRecord::IrreversibleMigration
+    # Define the reverse of this migration if needed
+    create_table :tasks do |t|
+      t.string :name
+      t.boolean :completed, default: false
+      t.integer :user_id
+      t.timestamps
+    end
+    add_index :tasks, :user_id
   end
 end
